test(color): migrate color picket spec from js to ts
