+++ tests/color/color-picker.spec.ts-snapshots/darwin/chrome/board-recent-color.png
