+++ tests/color/color-picker.spec.ts-snapshots/darwin/chrome/colors-panel.png
