+++ tests/color/color-picker.spec.ts-snapshots/darwin/chrome/rectangle-color-B1B2B5.png
