+++ tests/color/color-picker.spec.ts-snapshots/darwin/chrome/rectangle-color-FF0000.png
