+++ tests/color/color-picker.spec.ts-snapshots/darwin/chrome/rectangle-file-library-color.png
