+++ tests/color/color-picker.spec.ts-snapshots/darwin/firefox/board-recent-color.png
