+++ tests/color/color-picker.spec.ts-snapshots/darwin/firefox/colors-file-library.png
