+++ tests/color/color-picker.spec.ts-snapshots/darwin/firefox/colors-panel.png
