+++ tests/color/color-picker.spec.ts-snapshots/darwin/firefox/rectangle-color-B1B2B5.png
