+++ tests/color/color-picker.spec.ts-snapshots/darwin/firefox/rectangle-color-FF0000.png
