+++ tests/color/color-picker.spec.ts-snapshots/darwin/firefox/rectangle-file-library-color.png
