+++ tests/color/color-picker.spec.ts-snapshots/darwin/webkit/board-recent-color.png
