+++ tests/color/color-picker.spec.ts-snapshots/darwin/webkit/colors-panel.png
