+++ tests/color/color-picker.spec.ts-snapshots/darwin/webkit/rectangle-color-B1B2B5.png
