+++ tests/color/color-picker.spec.ts-snapshots/darwin/webkit/rectangle-color-FF0000.png
