+++ tests/color/color-picker.spec.ts-snapshots/darwin/webkit/rectangle-file-library-color.png
