+++ tests/color/color-picker.spec.ts-snapshots/linux/chrome/board-recent-color.png
