+++ tests/color/color-picker.spec.ts-snapshots/linux/chrome/colors-panel.png
