+++ tests/color/color-picker.spec.ts-snapshots/linux/chrome/rectangle-color-B1B2B5.png
