+++ tests/color/color-picker.spec.ts-snapshots/linux/chrome/rectangle-color-FF0000.png
