+++ tests/color/color-picker.spec.ts-snapshots/linux/chrome/rectangle-file-library-color.png
