+++ tests/color/color-picker.spec.ts-snapshots/linux/firefox/board-recent-color.png
