+++ tests/color/color-picker.spec.ts-snapshots/linux/firefox/colors-file-library.png
