+++ tests/color/color-picker.spec.ts-snapshots/linux/firefox/rectangle-color-B1B2B5.png
