+++ tests/color/color-picker.spec.ts-snapshots/linux/firefox/rectangle-color-FF0000.png
