+++ tests/color/color-picker.spec.ts-snapshots/linux/firefox/rectangle-file-library-color.png
