+++ tests/color/color-picker.spec.ts-snapshots/linux/webkit/board-recent-color.png
