+++ tests/color/color-picker.spec.ts-snapshots/linux/webkit/colors-file-library.png
