+++ tests/color/color-picker.spec.ts-snapshots/linux/webkit/colors-panel.png
