+++ tests/color/color-picker.spec.ts-snapshots/linux/webkit/rectangle-color-B1B2B5.png
