+++ tests/color/color-picker.spec.ts-snapshots/linux/webkit/rectangle-color-FF0000.png
