+++ tests/color/color-picker.spec.ts-snapshots/linux/webkit/rectangle-file-library-color.png
