+++ tests/color/color-picker.spec.ts-snapshots/win32/chrome/board-recent-color.png
