+++ tests/color/color-picker.spec.ts-snapshots/win32/chrome/colors-file-library.png
