+++ tests/color/color-picker.spec.ts-snapshots/win32/chrome/colors-panel.png
