+++ tests/color/color-picker.spec.ts-snapshots/win32/chrome/rectangle-color-B1B2B5.png
