+++ tests/color/color-picker.spec.ts-snapshots/win32/chrome/rectangle-color-FF0000.png
