+++ tests/color/color-picker.spec.ts-snapshots/win32/chrome/rectangle-file-library-color.png
