+++ tests/color/color-picker.spec.ts-snapshots/win32/firefox/board-recent-color.png
