+++ tests/color/color-picker.spec.ts-snapshots/win32/firefox/colors-panel.png
