+++ tests/color/color-picker.spec.ts-snapshots/win32/firefox/rectangle-color-B1B2B5.png
