+++ tests/color/color-picker.spec.ts-snapshots/win32/firefox/rectangle-color-FF0000.png
