+++ tests/color/color-picker.spec.ts-snapshots/win32/firefox/rectangle-file-library-color.png
