+++ tests/color/color-picker.spec.ts-snapshots/win32/webkit/board-recent-color.png
