+++ tests/color/color-picker.spec.ts-snapshots/win32/webkit/colors-file-library.png
