+++ tests/color/color-picker.spec.ts-snapshots/win32/webkit/rectangle-color-B1B2B5.png
