+++ tests/color/color-picker.spec.ts-snapshots/win32/webkit/rectangle-color-FF0000.png
