+++ tests/color/color-picker.spec.ts-snapshots/win32/webkit/rectangle-file-library-color.png
